backend/requirements.txt
fastapi == 0.85.1
uvicorn == 0.18.3
motor == 3.2.0
python-dotenv==1.0.0
mangum